requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0
mcp>=1.0.0
pandas>=2.0
scikit-learn>=1.3
boto3>=1.34
python-dotenv>=1.0
